+++ src/main/java/spinbox/commands/DoneCommand.java
@@ -3,8 +3,8 @@ package spinbox.commands;
 import spinbox.exceptions.SpinBoxException;
 import spinbox.exceptions.InputException;
 import spinbox.Storage;
-import spinbox.lists.TaskList;
-import spinbox.items.tasks.Task;
+import spinbox.containers.lists.TaskList;
+import spinbox.entities.items.tasks.Task;
 import spinbox.Ui;
 
 import java.util.ArrayList;
