+++ src/main/java/spinbox/commands/ListCommand.java
@@ -1,8 +1,8 @@
 package spinbox.commands;
 
 import spinbox.Storage;
-import spinbox.lists.TaskList;
-import spinbox.items.tasks.Task;
+import spinbox.containers.lists.TaskList;
+import spinbox.entities.items.tasks.Task;
 import spinbox.Ui;
 
 import java.util.ArrayList;
