+++ src/main/java/spinbox/commands/SnoozeCommand.java
@@ -2,13 +2,13 @@ package spinbox.commands;
 
 import spinbox.DateTime;
 import spinbox.Storage;
-import spinbox.lists.TaskList;
+import spinbox.containers.lists.TaskList;
 import spinbox.Ui;
 import spinbox.exceptions.SpinBoxException;
 import spinbox.exceptions.InputException;
-import spinbox.items.tasks.Deadline;
-import spinbox.items.tasks.Event;
-import spinbox.items.tasks.Task;
+import spinbox.entities.items.tasks.Deadline;
+import spinbox.entities.items.tasks.Event;
+import spinbox.entities.items.tasks.Task;
 
 import java.util.ArrayList;
 import java.util.Date;
