+++ src/main/java/spinbox/commands/ViewScheduleCommand.java
@@ -3,9 +3,9 @@ package spinbox.commands;
 import spinbox.DateTime;
 import spinbox.exceptions.InputException;
 import spinbox.Storage;
-import spinbox.items.tasks.Schedulable;
-import spinbox.lists.TaskList;
-import spinbox.items.tasks.Task;
+import spinbox.entities.items.tasks.Schedulable;
+import spinbox.containers.lists.TaskList;
+import spinbox.entities.items.tasks.Task;
 import spinbox.Ui;
 
 import java.util.ArrayList;
