+++ src/main/java/spinbox/entities/Notepad.java
@@ -1,5 +1,6 @@
-package spinbox;
+package spinbox.entities;
 
+import spinbox.Storage;
 import spinbox.exceptions.DataReadWriteException;
 import spinbox.exceptions.FileCreationException;
 import spinbox.exceptions.InvalidIndexException;
