+++ src/main/java/spinbox/entities/items/File.java
@@ -1,4 +1,4 @@
-package spinbox.items;
+package spinbox.entities.items;
 
 import spinbox.exceptions.CorruptedDataException;
