+++ src/main/java/spinbox/entities/items/tasks/Lecture.java
@@ -1,4 +1,4 @@
-package spinbox.items.tasks;
+package spinbox.entities.items.tasks;
 
 import spinbox.DateTime;
