+++ src/test/java/ReminderCommandTest.java
@@ -1,14 +1,5 @@
-import spinbox.DateTime;
-import spinbox.Storage;
-import spinbox.lists.TaskList;
-import spinbox.Ui;
-import spinbox.commands.ReminderCommand;
-import spinbox.exceptions.SpinBoxException;
-import spinbox.items.tasks.Deadline;
-import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import java.util.Date;
 
 public class ReminderCommandTest {
 //
