+++ src/test/java/ViewScheduleTest.java
@@ -1,8 +1,7 @@
 import spinbox.DateTime;
-import spinbox.items.tasks.Deadline;
-import spinbox.items.tasks.Event;
-import spinbox.items.tasks.Schedulable;
-import spinbox.items.tasks.Task;
+import spinbox.entities.items.tasks.Deadline;
+import spinbox.entities.items.tasks.Event;
+import spinbox.entities.items.tasks.Schedulable;
 
 import java.util.Date;
